requirements.txt
python-dotenv
matplotlib
pandas
python-daemon
configparser
autopep8
pycodestyle
pytest-mock
pyinotify

